% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.default}
\alias{fit.incidence2}
\title{Fit a regression model}
\usage{
fit(dat, model, ...)

\method{fit}{default}(dat, model, ...)

\method{fit}{incidence2}(dat, model = c("negbin", "poisson"), alpha = 0.05, ...)
}
\arguments{
\item{dat}{An \link[incidence2:incidence]{incidence2::incidence} object.}

\item{model}{The regression model to fit (can be "negbin" or "poisson").}

\item{...}{Not currently used.}

\item{alpha}{Value of alpha used to calculate confidence intervals; defaults
to 0.05 which corresponds to a 95\% confidence interval.}
}
\value{
An object of class \code{incidence2_fit}.
}
\description{
Fit a regression model
}
