% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\alias{growth_rate.default}
\alias{growth_rate.incidence2_fit}
\title{Calculate growth rate and doubling time}
\usage{
growth_rate(x, ...)

\method{growth_rate}{default}(x, ...)

\method{growth_rate}{incidence2_fit}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{The output of function \code{\link[=fit_curve.incidence2]{fit_curve.incidence2()}}.}

\item{...}{Not currently used.}

\item{alpha}{Value of alpha used to calculate confidence intervals; defaults
to 0.05 which corresponds to a 95\% confidence interval.}
}
\description{
Calculate growth rate and doubling time
}
