% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{i2extras-defunct}
\alias{i2extras-defunct}
\alias{fit_model}
\title{Defunct functions in package i2extras}
\usage{
fit_model(x, model, ...)
}
\description{
The functions or variables listed here are now defunct, i.e. no longer
available.
}
\details{
\itemize{
\item The \code{fit_model()} generic and associated functions were removed in version
0.2.0.
}
}
\keyword{internal}
