% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2extras-package.R
\docType{package}
\name{i2extras-package}
\alias{i2extras}
\alias{i2extras-package}
\title{i2extras: Functions to Work with 'incidence2' Objects}
\description{
Provides functions to work with 'incidence2' objects, including a simplified interface for trend fitting and peak estimation. This package is part of the RECON (\url{https://www.repidemicsconsortium.org/}) toolkit for outbreak analysis (<https://www.reconverse.org/).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.reconverse.org/i2extras/}
  \item Report bugs at \url{https://github.com/reconverse/i2extras/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})

Authors:
\itemize{
  \item Thibaut Jombart \email{thibautjombart@gmail.com}
}

}
\keyword{internal}
