% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{incidence}
\alias{incidence}
\title{Compute the incidence of events}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{incidence2}{\code{\link[incidence2]{incidence}}}
}}

