% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NA_counts_}
\alias{NA_counts_}
\title{Generate NAs of the right type for counts}
\usage{
NA_counts_(x)
}
\arguments{
\item{x}{A count vector.}
}
\value{
A \code{NA} of the type matching the input.
}
\description{
Counts can be of type integer or double. When setting up NAs to counts, this
needs to be reflected by using the right type of NAs. This function addresses
this need.
}
\author{
Thibaut
}
