% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okerr.R
\name{okerr}
\alias{okerr}
\alias{ok}
\alias{ok.default}
\alias{ok.incidence2_fit}
\alias{err}
\alias{err.default}
\alias{err.incidence2_fit}
\title{Error handling for incidence2_fit objects}
\usage{
ok(x, ...)

\method{ok}{default}(x, ...)

\method{ok}{incidence2_fit}(x, ...)

err(x, ...)

\method{err}{default}(x, ...)

\method{err}{incidence2_fit}(x, ...)
}
\arguments{
\item{x}{The output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, or
brms_model or a list of these objects.}

\item{...}{Additional arguments passed to underlying models.}
}
\description{
Error handling for incidence2_fit objects
}
