% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot.incidence2_fit}
\alias{plot.incidence2_fit}
\title{Plot a fitted epicurve}
\usage{
\method{plot}{incidence2_fit}(x, include_warnings = TRUE, ci = TRUE, pi = FALSE, ...)
}
\arguments{
\item{x}{An \code{incidence2_fit} object created by \code{\link[=fit_curve]{fit_curve()}}.}

\item{include_warnings}{Include results in plot that triggered warnings but not errors.

Defaults to \code{FALSE}.}

\item{ci}{Plot confidence intervals.

Defaults to TRUE.}

\item{pi}{Plot prediction intervals.

Defaults to FALSE.}

\item{...}{Additional arguments to be passed to \code{\link[incidence2:plot.incidence2]{incidence2::plot.incidence2()}} .}
}
\value{
An incidence plot with the addition of a fitted curve.
}
\description{
Plot a fitted epicurve
}
\author{
Tim Taylor
}
