% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.incidence2_ra}
\alias{plot.incidence2_ra}
\title{Plot a rolling average incidence object}
\usage{
\method{plot}{incidence2_ra}(x, ...)
}
\arguments{
\item{x}{An \code{incidence2_ra} object created by \code{\link[=add_rolling_average]{add_rolling_average()}}.}

\item{...}{Additional arguments to be passed to
\code{\link[incidence2:plot.incidence2]{incidence2::plot.incidence2()}} or \code{\link[incidence2:plot.incidence2]{incidence2::facet_plot()}}.}
}
\value{
An incidence plot with the addition of a rolling average.  This will
be facetted if the object is grouped.
}
\description{
Plot a rolling average incidence object
}
