% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rolling_average}
\alias{plot_rolling_average}
\title{Add a rolling average to an incidence plot}
\usage{
plot_rolling_average(graph, x)
}
\arguments{
\item{graph}{An incidence plot.}

\item{x}{An \code{incidence2_ra} object created by \code{\link[=add_rolling_average]{add_rolling_average()}}.}
}
\value{
An incidence plot with the addition of a rolling average.
}
\description{
Add a rolling average to an incidence plot
}
