% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_average.R
\name{add_rolling_average}
\alias{add_rolling_average}
\alias{add_rolling_average.default}
\alias{rolling_average.default}
\alias{add_rolling_average.incidence2}
\alias{rolling_average.incidence2}
\alias{remove_rolling_average}
\title{Add and remove a rolling average}
\usage{
add_rolling_average(x, ...)

\method{add_rolling_average}{default}(x, ...)

\method{add_rolling_average}{incidence2}(x, before = 2, ...)

remove_rolling_average(r)
}
\arguments{
\item{x}{An \link[incidence2:incidence]{incidence2::incidence} object.}

\item{...}{Not currently used.}

\item{before}{how many prior dates to group with.}

\item{r}{Am object generated by \link{add_rolling_average}.}
}
\value{
\itemize{
\item \code{rolling_average} An \link[incidence2:incidence]{incidence2::incidence} like object with an additional,
averaged, column and class \code{incidence2_ra}.
\item \code{remove_rolling} An \code{\link[=incidence]{incidence()}} object.
}
}
\description{
Add and remove a rolling average
}
\note{
If groups are present the average will be calculated across each
grouping, therefore care is required when plotting.
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE) &&
    requireNamespace("incidence2", quietly = TRUE)) {
  data(ebola_sim_clean, package = "outbreaks")
  dat <- ebola_sim_clean$linelist

  inci <- incidence2::incidence(dat,
                    date_index = date_of_onset,
                    interval = "week",
                    last_date = "2014-10-05",
                    groups = gender)

  ra <- add_rolling_average(inci, before = 2)
  inci \%>\%
    incidence2::facet_plot(color = "white") \%>\%
    plot_rolling_average(ra)


  inci2 <- incidence2::regroup(inci)
  ra2 <- add_rolling_average(inci2, before = 2)
  inci2 \%>\%
    plot(color = "white") \%>\%
    plot_rolling_average(ra2)

}

}
