% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.default}
\alias{fit_model.trending_model}
\title{Fit a model to an incidence object}
\usage{
fit_model(x, model, ...)

\method{fit_model}{default}(x, model, ...)

\method{fit_model}{trending_model}(x, model, ...)
}
\arguments{
\item{x}{An \link[incidence2:incidence]{incidence2::incidence} object.}

\item{model}{Either an individual \code{\link{trending_model}} object generated by
\code{lm_model}, \code{glm_model}, \code{glm_nb_model}, \code{brms_model} or a list of these
models (see \code{?trending::trending_model()} for further details).}

\item{...}{Not currently used.}

\item{ci}{Logical indicating whether to calculate confidence intervals for
the model (defaults to TRUE).}

\item{pi}{Logical indicating whether to calculate prediction intervals for
the model (defaults to TRUE).}

\item{alpha}{Value of alpha used to calculate confidence  / prediction
intervals; defaults to 0.05 which corresponds to a 95\% confidence interval.}
}
\value{
An object of class \code{incidence2_fit} or \code{incidence2_fit_list} if a
list of models is given as input.
}
\description{
Fit a model to an incidence object
}
\author{
Tim Taylor
}
