% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\alias{growth_rate.default}
\alias{growth_rate.incidence2_fit}
\title{Calculate growth/decay rate}
\usage{
growth_rate(x, ...)

\method{growth_rate}{default}(x, ...)

\method{growth_rate}{incidence2_fit}(
  x,
  alpha = 0.05,
  growth_decay_time = TRUE,
  include_warnings = FALSE,
  ...
)
}
\arguments{
\item{x}{The output of \code{fit_curve()}.}

\item{...}{Not currently used.}

\item{alpha}{Value of alpha used to calculate confidence intervals; defaults
to 0.05 which corresponds to a 95\% confidence interval.}

\item{growth_decay_time}{Should a doubling/halving time and corresponding
confidence intervals be added to the output. Default TRUE.}

\item{include_warnings}{Include models in output that triggered warnings but
not errors.  Defaults to \code{FALSE}.}
}
\description{
Calculate growth/decay rate
}
\author{
Tim Taylor
}
