% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\alias{add_growth_rate}
\alias{add_growth_rate.default}
\alias{growth_rate.default}
\alias{add_growth_rate.incidence2_fit}
\alias{growth_rate.incidence2_fit}
\title{Calculate growth/decay rate}
\usage{
add_growth_rate(x, ...)

\method{add_growth_rate}{default}(x, ...)

\method{add_growth_rate}{incidence2_fit}(x, alpha = 0.05, growth_decay_time = TRUE, ...)
}
\arguments{
\item{x}{The output of function \code{\link[=fit_curve.incidence2]{fit_curve.incidence2()}}.}

\item{...}{Not currently used.}

\item{alpha}{Value of alpha used to calculate confidence intervals; defaults
to 0.05 which corresponds to a 95\% confidence interval.}

\item{growth_decay_time}{Should a doubling/halving time and corresponding
confidence intervals be added to the outpu. Default TRUE.}
}
\description{
Calculate growth/decay rate
}
