% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okerr.R
\name{okerr}
\alias{okerr}
\alias{ok}
\alias{ok.default}
\alias{ok.incidence2_fit}
\alias{err}
\alias{err.default}
\alias{err.incidence2_fit}
\title{Error handling for incidence2_fit objects}
\usage{
ok(x, ...)

\method{ok}{default}(x, ...)

\method{ok}{incidence2_fit}(x, ...)

err(x, ...)

\method{err}{default}(x, ...)

\method{err}{incidence2_fit}(x, ...)
}
\arguments{
\item{x}{The output of function \code{\link[=fit_curve.incidence2]{fit_curve.incidence2()}}.}

\item{...}{Not currently used.}
}
\description{
These functions are used to filter succesful model fits from those that
errored.
}
\details{
The following accessors are available:
\itemize{
\item \code{ok()}: returns rows from an \code{\link{incidence2_fit}} object that did not error
during the model fitting stages.
\item \code{err()}: returns rows from an \code{\link{incidence2_fit}} object that errored
during the model fitting stages.
}
}
