% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot.incidence2_fit}
\alias{plot.incidence2_fit}
\title{Plot a rolling average incidence object}
\usage{
\method{plot}{incidence2_fit}(x, ...)
}
\arguments{
\item{x}{An \code{incidence2_fit} object created by \code{\link[=fit]{fit()}}.}

\item{...}{Additional arguments to be passed to
\code{\link[incidence2:plot.incidence2]{incidence2::plot.incidence2()}} or \code{\link[incidence2:plot.incidence2]{incidence2::facet_plot()}}.}
}
\value{
An incidence plot with the addition of a fitted curve.  This will
be facetted if the object is grouped.
}
\description{
Plot a rolling average incidence object
}
